moved getRequestTypeConfig to utilities
