+++ new-commands/utilities.js
@@ -197,3 +197,18 @@ export async function deleteThread(channelId) {
 
   return deleteResponse;
 }
+
+export async function getRequestTypeConfig(guildId, requestType) {
+  // Let's try to find the Label from requestType
+  const serverConfig = await readJSONFile(`./config/${guildId}.json`);
+  const requestHelpConfig = { ...serverConfig.requestTypes.find(i => i.value == requestType) };
+
+  const configFields = ['targetChannel', 'label', 'value', 'description', 'placeholder', 'emoji', 'colors', 'initialThreadMessage'];
+
+  const returnObject = {};
+  configFields.forEach(i => {
+    returnObject[i] = requestHelpConfig[i] || serverConfig.requestMeta[i];
+  });
+
+  return returnObject;
+}
